LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-rowspan-002-expected.txt

Visibility collapse

Spec
Setting a row to visibility:collapse changes table height but not width. Here, the spanning cell is the only cell in its row.

B
B
B
B
B
aaaa
bbbb
cccc
dddd
Bottom table is identical to top except second row has been collapsed. The spanning cell's height shrinks and is clipped.
B
B
B
B
B
aaaa
cccc
dddd
This text should not intersect with the table.

PASS spanning row visibility:collapse doesn't change table width
PASS fourth row stays the same height
PASS third row stays the same height
FAIL spanning row visibility:collapse makes row height 0 assert_equals: spanning row visibility:collapse makes row height 0 expected 0 but got 29
PASS first row stays the same height
FAIL spanning cell shrinks to sum of remaining three rows' height assert_equals: spanning cell shrinks to sum of remaining three rows' height expected 87 but got 116
